+++ core/src/test/scala/ai/lum/odinson/events/TestEvents.scala
@@ -1,9 +1,13 @@
-package ai.lum.odinson
+package ai.lum.odinson.events
 
 import ai.lum.odinson.utils.OdinResultsIterator
 import org.scalatest._
 
-class TestEvents extends FlatSpec with Matchers {
+import ai.lum.odinson.{EventMatch, OdinsonMatch}
+
+import ai.lum.odinson.BaseSpec
+
+class TestEvents extends BaseSpec {
 
   import TestEvents._
 
@@ -16,8 +20,8 @@ class TestEvents extends FlatSpec with Matchers {
   """
 
   // extractor engine persists across tests (hacky way)
-  val doc = Document.fromJson(json)
-  val ee = TestUtils.mkExtractorEngine(doc)
+  val doc = getDocumentFromJson(json)
+  val ee = Utils.mkExtractorEngine(doc)
 
   "Odinson" should "match event with promoted entities" in {
     val pattern = """