+++ CHANGES.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   be matched by the query. In other words, it skips the `MatchSelector`.
 - Added `buildinfo.json` file to the index to store versions and build info.
 ### Changed
+- Different organization for tests. Now every test extends a `BaseSpec` class and there are 6 categories of tests.
 - Turn `State` into a trait with very basic `SqlState` and even more basic `MemoryState` and placeholder `FileState` implementations by [@kwalcock](https://github.com/kwalcock)
 - REST API: `/api/parent` -> `/api/parent/by-document-id` & `/api/parent/by-sentence-id`
 - REST API: `sentId` param for `/api/sentence` -> `sentenceId`